@@ -1,5 +1,3 @@
-import org.springframework.boot.gradle.plugin.ResolveMainClassName
-
 plugins {
     id 'org.springframework.boot' version '2.4.1'
     id 'io.spring.dependency-management' version '1.0.9.RELEASE'
@@ -7,8 +5,8 @@ plugins {
 }
 
 group = 'serejka.telegram.bot'
-version = '0.0.1-SNAPSHOT'
-sourceCompatibility = '8'
+version = '1.0'
+sourceCompatibility = '15'
 
 repositories {
     mavenCentral()
